Primeros pasos creación API REST con JAVA y SpringBoot

+++ BackEnd/src/main/java/com/example/restaurant/Repository/MesaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MesaRepository extends JpaRepository<Mesa, Integer> {
+public interface MesaRepository extends JpaRepository<Mesa, Long> {
 
 }
